requirements.txt
#Core Requirements
picamera
imutils
pyzmq


#Test Requirements
nose
sphinx